Use Get/SetCreator for marking player ownership
I forgot about these functions, but in my defense there are a lot of
functions around and they didn't actually do the things I wanted them to
do in the first place.

However, I see in the next update a bunch more stuff is going to be
using them so I figured it'd be a good idea to do some transparent
mucking about so they work to my needs while still acting like they do
in the base game.