Add a working ammo menu
Lots of mucking about involved in this one, but at least I've now worked
out how to do the layout I want without having to resort to manual pixel
bashing! I'll have to backport it to the class menu.

I'm reasonably certain this is now safe in regards to the client and
server both using the same validation and people won't end up with weird
errors but I guess we can only figure that out with playtesting.