Make the exploit detection more reliable
The bookshelf model was getting deleted effectively at random and I
added the visual warning to track it down - turns out the collision
model is offset and you can sink it very slightly into the world.

I went for always using the prop's centre because it's pleasantly
predictable and means that the player can now only ever sink 49% of the
model into the world, while previously if they flipped a prop upsidown
they could get 99% of it in there.

I left the warning in place because while it would make abusing any
theoretical exploit much easier, it also removes any confusion an
innocent player would have if a bunch of their props disappeared without
warning when they put them in the wrong place.

FWIW I don't actually know what exploit this code was originally added
to the gamemode to thwart - I really can't think of anything off the top
of my head (certainly not with the new prop rules) but you kinda don't
want to delete code explicitly marked as an exploit fix do you?