Super shotgun!
I started with this on the basis that it'd be easy, and then it turned
out to possibly be one of the most complicated guns in the game.

I could have got away with not trying to mimic the base game's shotgun
so perfectly, but I couldn't resist the challenge. This is actually a
bit of a mashup of the hl2 and hl2mp shotguns because they do actually
behave differently and I didn't like how the hl2mp one lets you skip
animations if you do the right sequence of buttons.

I'm ... suprised to say the least about the force it sends you back by,
shooting straight down will send you high enough to get fall damage and
jumping and firing at an angle will let you cross the map in a single
glorious bound - but it's been like that so who am I to quibble.

I did remove the 15 degree aim check because it seemed extremely
arbitrary, if we end up adding it back it should be a gradual lessening
and absolutely not happpen when you're in the air!