Scale onslaught damage based on the skill cvar
While mucking about in the SDK looking for something else, I discovered
that hl2 (and gmod!) scales damage based on the `skill` cvar, so in
order to make it possible for server owners to quickly tune the
difficulty of the server (and me to make it easier for testing) I've
decided to go all-in on support.

The only bit of this that was hard was getting the bullet damage for the
super shotty to work properly, but I discovered that setting the damage
to 0 causes the game to look it up from the ammo and then scale it
appropriately, so here we are.